+++ resources/src/js/main.js
@@ -1,6 +1,8 @@
 // load components
 Vue.component("leaderboard", require("./components/leaderboard/LeaderboardComponent.vue").default)
-Vue.component("scored-entry-prediction", require("./components/leaderboard/ScoredEntryPredictionComponent.vue").default)
+Vue.component("leaderboard-page", require("./components/leaderboard/LeaderboardPageComponent.vue").default)
+Vue.component("round-navigation", require("./components/leaderboard/RoundNavigationComponent.vue").default)
+Vue.component("scored-entry", require("./components/leaderboard/ScoredEntryComponent.vue").default)
 
 Vue.component("registration-workflow", require("./components/registration/RegistrationWorkflowComponent.vue").default)
 Vue.component("registration-entry", require("./components/registration/RegistrationEntryComponent.vue").default)